requirements.txt
markdown==2.6.8
pelican==3.7.1
pyyaml>=4.2b1
